requirements.txt
book-club==0.1
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
requests==2.7.0
